requirements.txt
amqp==2.6.0
appdirs==1.4.3
asgiref==3.2.8
aspy.refactor-imports==2.1.1
astroid==2.4.2
backports.zoneinfo==0.2.1
billiard==3.6.3.0
black==20.8b1
blessings==1.7
boto3==1.14.4
botocore==1.17.4
bpython==0.21
CacheControl==0.12.6
cached-property==1.5.2
celery==4.4.7
certifi==2021.10.8
cffi==1.14.4
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.3.1
curtsies==0.3.5
cwcwidth==0.1.4
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
distlib==0.3.1
distro==1.4.0
dj-rest-auth==2.1.2
Django==3.0.7
django-allauth==0.44.0
django-celery-beat==2.0.0
django-cors-headers==3.4.0
django-debug-toolbar==3.1.1
django-extensions==3.1.1
django-model-utils==4.0.0
django-naomi==0.8
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-revproxy==0.10.0
django-simple-history==2.11.0
django-taggit==1.3.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.1.1
filelock==3.0.12
flake8==3.8.3
flower==0.9.5
future==0.18.2
geocoder==1.38.1
greenlet==1.0.0
html5lib==1.0.1
humanize==3.0.1
identify==1.5.5
idna==2.9
importlib-metadata==1.6.1
importlib-resources==3.0.0
ipaddr==2.2.0
isort==4.3.21
jmespath==0.10.0
klepto==0.1.9
kombu==4.6.10
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
mypy-extensions==0.4.3
nodeenv==1.5.0
numpy==1.19.2
oauthlib==3.1.0
packaging==20.3
pathspec==0.8.1
pep517==0.8.2
phonenumbers==8.12.21
Pillow==7.2.0
pox==0.2.8
pre-commit==2.7.1
progress==1.5
prometheus-client==0.8.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytoml==0.1.21
pytz==2020.1
pyxdg==0.27
PyYAML==5.3.1
ratelim==0.1.6
redis==3.5.3
regex==2020.11.13
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
sentry-sdk==1.5.6
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
timezonefinder==4.4.1
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3
vine==1.3.0
virtualenv==20.0.31
webencodings==0.5.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
